Suppress known warnings on command line builds
We develop Java in an IDE, where we get more warnings than I can shake a
stick at, so there is no real need to have Java warnings spew out when
doing a command line build. [The case in question was a semantic
warning from Jikes about certain package classes not being in their
own files. Whatever]

JNI side, at the moment we are casting jlongs to C pointers on the way
in and casting C pointers into jlongs on return. This results in a
warning about pointer to wrong-sized-int converstion. That's a known
issue: a) so far so good, it seems to work fine right now; even so, b)
we may deal with it in the future with a macro or function wrapper. In
the mean time, the warning is distracting.