Delegate memory management for Boxed to individual subclasses
We know the copy and free functions for each Boxed from the defs
files, end most of the time there is a function or method listed with
each such type called _free() anyway.

So rather than trying to store the GType on the Java side somewhere in
order to be able to call the hopefully registered free function
through some all encompassing free in GBoxed.c, have each Proxy
sublcass implement a release() function [called by finalizer
presumably] which can Do The Right Thing (tm), and try this for Boxed.
Add a brief mockup for TreeIter to model what this would look like for
such a Boxed class on the Java side.

Boxed adds an boolean which will cause us to call into the free only
if we are the owner. Default this to true for the moment. We'll have
to work out how to act on the exceptions to this later.

Note that Boxed does NOT have a default fallback to calling g_free()!