Initial experiment with thread safety
Calling Gtk.mainQuit() from a Thread caused the VM to segfault, which
is not surprising since I hadn't done anything sepecial to get that
call back into the main execution thread.

On a hunch, added the code to Gtk.c to initialize the GThread system
and to tell GDK to be threaded. Then putting the gdk_threads_enter()
and gdk_threads_leave() guards around the call to gtk_main_quit()
caused things to run fine. This seems to be an indication that we may
well be able to be thread safe without any further effort than using
the guards.