Major import: signal handling
Add connectSignal() to org.gnome.glib.Plumbing, which in turn calls
org.gnome.glib.GObject's g_signal_connect() with the Object being
connected to, an instance of a Signal subclass, the Class containing
the handleXXX() callback, and the name of the signal.

Implement g_signal_connect() in org_gnome_gtk_GObject.c, 
Using the signal name, the signal is looked up, a GClosure created,
and then the signal is connected with our new marshaller.

To form the necessary GClosure and to do the signal marshalling we add
bindings_java_signal.c containing this code. There are two signficant
entry points:

bindings_java_closure_new() is where we populate our custom GClosure
and return it. It does the legwork to do all the JNI lookups
necessary. The most critical part is figuring out the JNI signature of
the appropriate (presumably generated) handleXXX() method. This is
done with the aid of the information provided by introspecting the
signal we're building a closure for, and is used to find the method ID
needed by the marshaller to invoke the method. All this information is
then stored in our custom GClosure struct extension called
BindingsJavaClosure.

bindings_java_marshaller() is the other significant entry point,
though it is called by GSignal upon a signal emission, rather than the
Java programmer. This function is likewise full of voodoo code. It
takes the parameters passed and converts each one into the appropriate
type based on our defined public/translation/native/jni/C argument
passing archiecture. Then, based on the return type identified and
stored by bindings_java_closure_new(), it calls the appropriate
JNI function to invoke the handleXXX() method on the static class on the
in the translation layer, which takes care of invoking the callback in
the handler instance from there. [And a better comment in Signal
indicating how this works].

bindings_java.h updated accordingly, with comments to indicate which
source file is the origin of which functions.