Demonstrate handling Window closing by hooking up Window.DELETE
The signal emitted when a window is closed is "delete-event", however,
since this is a GdkEvent bubbling up, it is handled by GTK in
GtkWidget. Therefore the generated code [will] have an interface called
GtkWidget.DELETE_EVENT extends Signal. This, however, is a
non-intuitive place for it, seeing as how this signal only ever
arrises on GtkWindows.

So, demonstrating that our public API can make choices to change the
mapping when appropriate, add signal interface Window.DELETE extends
GtkWidget.DELETE_EVENT.

While we're at it, introduce significant JavaDoc documentation on the
signal and on setTitle() and setDecorated().

Had to revert the gdk_threads_enter()/leave() in the implementation of
Gtk.gtk_main_quit(). Although it worked when calling from another
thread case, it caused a regression when working from the main thread.
Clearly that needs more work.