Initial import: prototype code generator
This is the initial import of the Perl program which generates the
java-gnome bindings from a series of .defs files.
  
This is fairly cumbersome code. For one thing the .defs file parser
and the code generator are all jammed together, one calling the other
at the conclusion of each block. However, it does demonstrate some of
the basic idioms that will be necessary, and validates the idea of
using the C types as the hash keys for the various layer lookups.