Ensure Makefile refuses to run if .config is not present or up-to-date
Simple make target depending on .config; if it is not present then
./configure needs to be run, so output an appropriate error message
and then abort.

Somehow this got left out when we imported Equivalence from another
project; copy and paste directly from slashtime, changing the file we
depend on for the version constant to src/java/org/gnome/gtk/Gtk.java.