Correct licence headers
Add top level LICENCE file, using the one from xseq, objective and
slashtime as a starting point and updating it with the text of the
"Classpath Exception".

Good form has it that each source file contains, at its start, a
header comment describing the licence that that file is released
under. 

Personally, I see no reason whatsoever for the entire text of the GPL
to be present in each file, so these header comments simply refer back
to the top level LICENCE file. They leave no doubt, however, that this
code is made available by the copyright holders under the terms of the
GPL.

While correcting the licence header I ended up touching every single
source file, so there are various minor textual changes here and there
- mostly just adding missing private constructors.

The dummy mockup of GnomeEntry was completly in the wrong tree. Fixed.