Add MARKUP file
Create a new file containing documentation of the Markdown markup
syntax used in our documentation.

The brief overview describes italics, bold, constant width, code
blocks and headings.

Include reference to where Markdown can be downloaded from (both Perl
and PHP) and that it is used in the java-gnome website to render our
docs.