Add copyright+licence statement to documentation files.
Following the good form of having a copyright header noting the GPL
licence on a per file basis, we add an appropriate statement to each
of our Documentation files. Such a comment, however, looks like shit
at the top of a text file especially when these text files are
intended to be easily readable from a terminal. So, {shrug}, put the
copyright statement in a footer comment at the bottom. Incorporate the
vim modeline into it for cleanliness.

Add relative hyperlinks from HACKING to files in doc/style and
doc/design.

Move MARKUP file to doc/style. Adjust references accordingly.