Add relative references from README
In the opening section of README, itemize each of the "Re-engineering"
documents as now contained in doc/design/, and link to them. a listing
by filename or an automatic directory index would do as the files are
reasonably well named, but proper hyperlinks with more descriptive
text are suprior, of course.