Update Equivalence with files from slashtime
The Equivalence setup originally imported from my private libgtk-java
had all kinds of Gentoo specific hardwired paths and was never updated
for Debian and Fedora. So fix that by merging in a more recent vintage
configure and Makefile from the [java-gnome 2.x] program `slashtime`.

John Rice <john.rice@sun.com> with Sun Ireland was kind enough to
offer fixes to the Equivalence setup for slashtime so the above merge
should enable the java-gnome 4.0 prototype to build on Solaris as
well.