Fix JNI includes for Linux
The logic for detecting which include paths to use as -I arguments to
teh C compiler was a bit broken for GCC on Linux for the case of GCJ
not installed.

Change to -I BOTH $java_home/include and $java_home/include/linux.

Thanks to Srichand Pendyala <srichand.pendyala@gmail.com> for the
insight that led to this fix.