Change relative references to target .html
At various points in the documentation we use Markdown's linking
mechanism to point at other files. Change the references to point not
at the text file but at the same filename with a .html suffix. This
will generate a 404, which will cause render.php to use Markdown to
process the .txt file into an HTML stream.