Fix JAVADOC invokation
A typo in the top level configure script resulted in a misnamed
variable being output to .config and the wrong value being placed in
$JAVADOC. Fixed.

Have the `make doc` target depend on a successful java -> class
compilation as a syntax check before unleashing the not-as-powerful
javadoc parser on it.

Minor improvements to the Equivalence output in silent mode.

Have Bazaar ignore the generated output in doc/api/ if present.