Add make dist target
Introduce new variable VERSION to org.gnome.gtk.Gtk, being the current
release version of the library. Adjust the top level configure to
parse that in additon to APIVERSION and output it to .config

Use VERSION to form the name of the tarball produced when the `make
dist` target is called. This target depends on a complete build as an
obvious preparedness verification. Then as a sanity check make sure
there are no uncommitted (especially seeing as how one of those
changes would likely be the version number).

Import the (missing) sanity check to make sure .config was sourced
properly.

Improve thouroughness of `make clean` and `make distclean`.