Cosmetic: Adjust code examples to a single 4 space indent
Rather than having to bother about 4 spaces (or worse one tab) to turn
on preformatting, and a second set of 4 spaces to shift the code right
by a margin, instead use the CSS stylesheet to do the margin indent.
So remove the double tab or 8 space indent from documentation files
and replace with 4 spaces.