Add a "favicon" for the website
Working from Joao Victor's original SVG of the java-gnome logo, create
a 16x16 pixel PNG and have the template link to it so that it appears
in tabs and in address fields in web browers.

It is a known bug that Internet Explorer only parses these files in
ICO format. Do we care?