Add sourceforge logo to template
Per the terms of service required by SourceForge, include their logo
in every page by putting it in the footer after the copyright message.

Import a copy of the logo, cropping out the black border.

Rename the image from OpenOffice to match the naming scheme used in
the other website images.