Add top level README file
At long last, add a top level README file to the project. Include
discussion of how to get the source code, how to build it, and,
briefly, what to do with it.

The discussion of how to specify options to the top level Equivalence
configure script is the text used in the HACKING file from Slashtime,
ObjectiveAccounts and other java-gnome programs written by Operational
Dynamics.