Cosmetic changes
Touch up the italics note text at the top of /doc/ and /get/.

Fix the background url for the separator image used to provide visual
distinctiveness between the maincontent and bottompanel areas.

Add some vertical whitespace on the home page below the main content
to reinforce its importance and to push the copyright babble further
down.