Introduce instanceFor() as an "abstract" method on Plumbing
Add new method instanceFor() to org.freedesktop.bindings.Plumbing,
replacing and absorbing proxyFor(). Since the method is static it
provides a skeleton method body, and and that's where we put the code
that used to be in proxyFor().

Nevertheless, the instanceFor() in org.freedesktop.bindings.Plumbing
should be considered abstract in the sense that any library using this
infrastructure will need to extend its own instanceFor() that knows
how to create Proxy instances for pointers that were not originally
created on the Java side.

Remove individual generated classes' instanceFor() methods, replacing
them with the instanceFor() method visible on org.gnome.glib.Plumbing,
and adding the necessary cast.