Implement getProperty() and g_object_get_property()
Implement the JNI code for g_object_get_property() which underlies
org.gnome.glib.Object's getProperty(). Add getPropertyString() as an
example of getting primative values out of properties.

This relies on the ability of Plumbing's instanceFor() to create
Proxies, which is still only in scratch form. Add a protected
constructor to StringValue to permit that to work.

In order to create the empty GValue that is to be passed to
g_object_get_property() use, g_object_class_find_property(), the idea
for which is borrowed from java-gnome 2.x's glib-java.