Replace g_value_type() with g_type_name()
The GType itself is rather useless; not only is it an opaque quantity
Java side, but worse they change each execution so switching on them
is pretty much useless, without constructing a map; we can do that far
more easily with Strings and the GType's name.

Add type() to GValue, but not sure if we need such a thing hacker
visible. Leave it there for now.