Replace instanceFor() with a reflexive version
Add a system whereby each of our Proxy types registers its GType name.
We actually do this in the static initializer of the generated
translation layer class (saving the effort of writing it by hand and
since the code generator has this information at hand). When the type
is registered it looks up the <init>(long) constructor which is used
to create that kind of Proxy subclass given a pointer.

Upgrade the instanceFor() method in org.gnome.glib.Plumbing to use
this table.  If a Proxy does not already exist for that pointer, then
an attempt will be made to construct one.

Relies on a protected <init>(long) being available in every Proxy
subclass, of course, so add one to each of the types in
Fundamental.java