Rename org.freedesktop.bindings.Enum to Constant
Following the best practice documented in _Hardcore Java_, include a
the name of the enum along with the ordinal when constructing the
constant objects that are our Enums subclasses.

The name Enum collides with a (somewhat internal) class in Java 1.5,
so rename it to Constant which is actually more accurate given the
constant object behaviour.

Fix WindowType.java to use new constructor signature.