Register Constants so they can be looked up by Plumbing
Add a knownConstants table to Plumbing. This mirrors the knownProxies
table in that it allows lookups of already instantiated objects.
Rather than pointer to Proxy, This one is a two tier table where
lookups are done by Class then indexed into an ArrayList resulting in
a Constant.

The top level constructor in Constant now makes a call to
Plumbing.registerConstant() to add the Constant object to the
knownConstants table.

Add constantFor() to Plumbing which returns the Constant corresponding
to a given Java Class and ordinal. This mirrors intanceFor() which
returns Proxy objects.

Considerable improvement to the JavaDoc and code comments throughout
this class.

Adjust the hash table types to no longer be WeakHashMap. That was
incorrect. In the case of knownProxies we need weak values, not keys,
and in the case of knownProxies it is not to use weak references on
either side.