Revamp org.gnome.glib's Plumbing to handle Constants and Enum[Values]
Considerable improvements to the Plumbing class responsible for
handling the GLib/GObject type system.

Add a special org.gnome.glib.EnumValue to Fundamental.java, usually
with a constructor that takes (long, Class) rather than just (long).

Add a mapping for types that happen to be enums from GType [name] to
the subclass of Constant that represents them. These values are
stored in an improved registerType() which uses further reflection to
introspect the passed classname. By having null in the original
typeMapping table, we signal that enumMapping, along with EnumValue()
should be used instead.

This work of looking up Proxies is still done by instanceFor(), which
has been significantly upgraded to instantiate Proxy subclasses for
normal Proxy (pointer) objects, but to use EnumValue when
instanceFor() is called on something that is actually a plain enum -
it means that the enum is wrapped in a GValue of G_TYPE_ENUM.