Introduce getPropertyEnum()
Add getPropertyEnum() as another protected method on
org.gnome.glib.Object. Returns the instance of a subclass of Constant
that corresponds to this particular Enum value.

We use Object's general getProperty(), expecting that a EnumValue will
be created and returned. In the translation layer, we use new method
GValue.getEnum() which calls new native method g_value_get_enum() to
extract the ordinal corresponding to the enum from the GValue which
wraps it. It then looks up the Constant with Plumbing.constantFor()
from org.freedesktop.bindings, and returns it.