Add @since tags
Because java-gnome is rapidly evolving, it is important to identify
when a given class or method appeared, so that someone perusing the
public API documentation who, for whatever reason, does not have the
latest code, is not completely left hanging. So add @since tags to all
public classes and public methods.

Do a library wide code format.