Place type mapping in external file
The idea of calling registerType() in the static initializer of each
generated translation class was unworkable because we cannot count on
the class having been loaded before we need to be able to map its
values (loading the class for the first time requires knowing which
class to load, which requires the mapping, so it can't be in that
class).

So move the mapping information to an external properties file. Add
typeMapping.properties to mockup/java/. This will be written by the
code generator. org.gnome.glib.Plumbing now loads this file from the
program's classpath via the getResource() capability in the
ClassLoader.

Remove the registerType() calls from the static blocks of the mocked
up translation classes.