Website sylesheet tweaks
Remove the style block in render.php that was setting Times New Roman
as the body font. Serifs are nice, but the rest of the site is in Sans
for the usual web screen usability reason, so the generated
documentation should likewise be in a Sans font for consistency.

Change the CSS for <pre> tags to be white on black, providing greater
visual distinction reminiscent of what would be seen on a terminal
screen.

Add an (underlined) link to "README" from the first paragraph of
/4.0/doc/ to be consistent with later links.

Use an SGML comment to nicely obscure the reference to the website at
the beginning of README