Add getURI() to FileChooser
Add getURI() to the FileChooser interface, mapping GTK's
gtk_file_chooser_button_get_uri(). Implement in FileChooserButton, but
go to the trouble of creating and returning a java.net.URI rather than
just returning a String - there's getFilename() for that.

Update the typeMapping properties file with the FileChooserButton and
FileChooserAction which were left out when those classes were added.