Add new public factory class Glade
The algorithmically mapped name Xml is just too harsh for the purpose
it serves. So, keeping with the practice in the other libraries, we
create a class named for the library with static methods used as
primary entry points.

Expose Glade.parse() wrapping Xml(String, String), which we now reduce
to package visibility.