Fix crasher when getting GType name
We rely on being able to retrieve the name that was registered for a
type corresponding to an arbitrary pointer representing some G struct
or another. This turns out to be somewhat tricky.

This patch fixes a crasher by a) explicitly casting a bunch of
pointers to be the correct type for each of a variety of checks, and
b) calling a more appropriate macro depending on the case.

Return an empty string to Java rather than null if the name is NULL so
that the call to .intern() doesn't result in a NullPointerException