Introduce new code path for GValues distinct from GObject
Mistakenly representing GObject as a subclass case of GValue for
property setting was something we were getting away with in some tests
but which eventually led to segfaults whenver trying to retrieve
objects once the instanceFor() to g_type_name() system was introduced.

So refactor org.gnome.glib.Object to be a direct subclass of
org.freedestkop.bindings.Proxy in parallel with org.gnome.glib.Value.
Create new Fundamental type ObjectValue for GValues holding GObjects.
Add new methods to org.gnome.glib.Object for the getPropertyObject()
case, which is a GValue which wraps a GObject.

Create a new GObject.g_type_name() equivalent to GValue.g_type_name()
but for the case of a GObject rather than a GValue. This is called by
instanceFor(), the default case used in generated code. Introduce new
method valueFor() in org.gnome.glib.Plumbing for use by the
getProperty() family of methods to extract the type of something known
to be a GValue (rather than a GObject).

Thanks to Davyd Madeley <davyd@madeley.id.au> for extensive debugging
assistance, and credit to Manish Singh <manish@flock.com>, James
Henstridge <james@jamesh.id.au>, and Malcolm Tredinnick
<malcolm@pointy-stick.com> for having analyzed the root cause issue
and having clarified that two code paths would indeed be necessary.