Temporarily use a JNI call to create Proxy instances
Since protected methods in other packages aren't visible to
org.gnome.glib.Plumbing, we can't use newInstance(), we need to use
the JNI NewObject() function as a workaround. So add Plumbing.c
implementing createInstance().

We'll replace this with a series of Factory classes which can register
themselves with Plumbing to keep in on the Java side in due course.