Support null argument for root
If the second argument to glade_xml_new() is NULL, then it creates
Widgets for all the hierarchies in the file. Alter
org.gnome.glade.Glade.parse(), Xml's <init>(String, String) and the
implementing JNI file to handle (pass through) the case of a Java null
being used for the second argument.

Spell check the public API documentation.