Fix codepaths for GValues containing GObjects
The returned type name of a GValue holding a GObject that happens to
be expected to be of type GtkContainer (even though the actuall object
is a GtkWindow or some such) is "GtkContainer", which doesn't help
much as what we actually need is "GValue<GObject>" or something along
those lines.

So, in the Value codepath used in property getters, hardwire an escape
clause into the case where the GValue happens to be holding a GObject,
and deal with that result in org.gnome.glib.Plumbing.