Introduce preliminary unit tests
Create unit tests for some of the basic characteristic getters and
setters that were being evaluated with asserts in Experiment. This
isn't the final resting place or layout for the tests, but we have to
start somewhere.

Create a top leve UnitTests class that contains a suite of all the
tests to be run (by referencing each test class by literal). Have it
call Gtk.init(). We'll need to refactor that, too.