Add make test target, and detect JUnit in configure
As imported, the Equivalence configure script looks for prerequisite
.jar files and forms a classpath from them. We had no need for that,
so it was all commented out or removed.

So restore (by copying in from ObjectiveAccounts) the prerequisite
check for JUnit's junit.jar and write it to .config as JUNIT_JARS

Add (re-enable) a test target in the Makefile, dittoing the make demo
target's execute, to call the main() function in class UnitTests and
utilizing the JUnit .jar as detected above.

Test classes are built to tmp/tests rather than tmp/classes so they
aren't packed up into the gtk-4.0.jar. Fix the Eclipse build settings
to correspond to this.