Import generic.* Text and Test utilities from ObjectiveAccounts
The UnitTests top level class we brought in from ObjectiveAccounts had
a main method which used a fancy JUnit wrapper classes to produce more
interesting output when the unit tests are running from in a text
console.

The pretty printer depended on a series of other utility classes
tucked away in packages generic.{ui,util,junit}. So, import the lot,
but since this is a public library, reparent the packages to
com.operationaldynamcis.{ui,util,junit}.

Other than formatting, this code hasn't been cleaned up. In
particular, the Environment wrapper can now be implemented to actually
get the value of an environment variable (rather than just a property)
by using a native method.