Realize make install target
Activate install as a target within the Makefile. Files are sent to
FHS compliant locations under $PREFIX, with $DESTDIR prepended so that
distros constructing binary packages can relocate the installed tree
relative to a temporary location.

Add prefix as an option to configure, and yes, support --prefix as
well, or people will be just totally futzed.

While we're at it, adjust the formatting (comment width) of the text
output to the .config fragment.