Detect Eclipse's standalone Java compiler
Add ecj to the list of compilers we look for. At the moment, this is
done properly by hard coded path on Gentoo, and by path lookup for
"ecj" on the other distros.

Bump ecj to #1 in the compiler selection order list. 

Add -g debug flag to all java compiler lines so that when step
debugging in an IDE you get actual variable names and not arg0.