Add g_object_ref to org.gnome.glib.GObject
Turns out we _do_ need to call g_object_ref(), in the case where we
have created a Proxy on our own, wrapping an object in a signal
handler or getter. In that case (as opposed to us calling the
constructor) we need to up the ref count. Internal to org.gnome.glib
only.