Fix Eclipse build configuration so CDT outline works
When we refactored from having the C code in src/jni | mockup/jni to
being alongside the Java classes whose native methods they implent, we
needed to change the path settings in the CDT part of the project
configuration. Now the outline function in C/C++ perspective works
again. Yeay.