Refactor instanceFor() to objectFor(); increase ref count
Refactor signature of instanceFor() in org.gnome.glib.Plumbing to
objectFor(). The original intent of having instanceFor() override the
static method provided on org.freedesktop.bindings.Plumbing turns out
to be a bit silly in the face of needing the valueFor() code path
special to GValues. So treat the GObject code path as being equally
special; when we generate the translation layer we will *know* what
type it is, and so can select the appropriate "instanceFor()"
implementation without needing to depend on it figuring it out
magically.

When we create a Proxy for a GObject pointer in response to a getter,
we need to manually increment the reference count with g_object_ref().
Do so in new private method createObject().