Add section about proper identification
Commit messages are important, which is why we have an entry in the
style guide about them. Since they are listed by first name and
committer name and address, its important that those points be right.

Add a section to CommitMessages.txt to remind people that they need to
set an appropriate identification via one of .bazaar/bazaar.conf, `bzr
whoami`, or BZR_EMAIL if they wish to have their patches accepted.