Add -sourcepath arguments to JAVAC lines.
It would seem that ecj has an issue when multiple top level classes
are in a single source file, but I may just be chasing my own tail.

Based on the documentation for javac, the present complilation line
seems to be the correct approach, and resolves a glitch that javac and
jikes were having. If ecj still has trouble, then we'll figure it out
later. Note that Eclipse itself doesn't have the problem, just
standalone ecj.