Adjust Makefile to export COLUMNS
The COLUMNS pseudo environment variable is not exported by bash. By
its own internal rules, Make doesn't see this variable, and so in turn
does not export it to child processes, which means that it isn't
available to the VerboseResultPrinter or TextOutput and so they can't
figure out how wide the screen is. How annoying.

After a bit of faffing about, change the logic in Makefile to parse
`stty size` and manually export it in a variable named COLUMNS. As
this is only necessary when running from console, and you're being a
good little boy or girl and running the test suite via `make test`,
you're going to have the correct value picked up and your output will
look all pretty.