Update README file to reflect installed locations
Move the bulk of the `make install` section from NEWS to README, which
is a better place for it. Further describe how to use java-gnome
(installed or not) to run your programs.

Apparently people didn't get the 4.0.7 joke. {sigh}. So change README
to say "get java-gnome-4.0.2.tar.whatver". That would seem to imply
we'll have to change that number each time there's a release. Ok,
fine, but I won't sweat if we forget once in a while.

Continue to tighten the prose in the NEWS file leading up to the 4.0.2
release.