Implement Container's remove().
In order to exercise whether or not we're interacting with the GObject
reference system properly, we needed to be able to remove a Widget
from a Container. So, add the method.

Significantly fix up Container's JavaDoc, adding the beginnings of a
class description and explaination for both add() and remove().