Use WeakReferences.
I thought it high time to check that calls to things like
GValue.free() were not segfaulting, only to realize that it wasn't
getting called at all because we hadn't gotten around to actually
finishing using weak references to the Proxies in Plumbing. Oops.

Of course, we'd been putting that off because of the complexity of
testing it and because we didn't want to chase null pointer bugs at
the time. So, deep breath:

Convert the values put into org.freedesktop.bindings.Plumbing's
knownProxies Map from the Proxy itself to WeakReferences to the Proxy,
and adjusting the getter code to unwrap accordingly.