Fix use of GObject refs
Move the debug output statement to release() rather than finalize()
since the latter may only be temporary and release() will be called
regardless.

Introduce new class org.freedesktop.bindings.Debug to contain global
conditional compilation constants. For now the only one is
MEMORY_MANAGEMENT, and use it to wrap the output statements in Object
and Value's release() methods.

Fix a major error in GObject reference handling. For some reason, I
had assumed that the floating reference created with _new() functions
was "ours" and that the sinking of that reference into a parent
container would magically work for us. This turns out to be a
ridiculous mistake: those references are held by the container and
when that container is destroyed the references to to the children are
automatically dropped (see the GNOME API documentation for the
underlying gtk_container_remove() for further discussion). The result
was that by the time g_object_unref() was called by our Object's
release() method, the GObject was already destroyed.  Ga-boom (though
not segfault, at least).

The solution was sorta obvious, and in hindsight is what should have
been done since the very beginning: call g_object_ref() when we make a
new Object extends Proxy, and we have our reference just like we
should have had all along.

It's tempting to move the g_object_ref() call to the constructor on
the JNI side to eliminate the round trip from Object.<init>(long).  We
can experiement with that later; admittedly doing so would break
encapsulation.