Add test cases to validate resource releasing
Alter VerboseResultPrinter so that when Debug.MEMORY_MANAGEMENT is set
it does not leave the cursor position at the right of the screen. That
way when debug statements in release() methods output to terminal they
don't look completely terrible.

Add unit test TestReferenceCounting. It's not much of a validation of
the memory management aspects of our design, but it did allow us to
find the major bug discussed in the previous commit.

Alter tests/prototype examples slightly to use a private constructor
so that the objects they create become unreachable within the running
program and eligible for finalization. This allows us to observe those
objects being released.