Rename test cases to Validate*
Using Test* as the prefix for the unit test class names turns out to
be little help; typing Test<COMPLETE> shows a zillion classes from
JUnit which are no help at all when trying to find one of our own test
cases.

So rename them to Validate*. It's not the most perfect name we could
have found, but it's a clean namespace.

Minor output touch ups to VerboseResultPrinter for debugging and
console output of JUnit failures or errors.