Move version and apiversion data to new class org.gnome.gtk.Version
We had the Equivalence setup depending on and extracting its data from
org.gnome.gtk.Gtk. Since we require a re-configure whenever the file
with the version constants in it changes, that made hacking on
Gtk.java a bit of a pain.

Move the APIVERSION and APIVERSION constants to new package visible
class org.gnome.gtk.Version. Adjust configure and Makefile
accordingly.