Override toString() in Proxy
Beginnings of a stab at doing a better toString() for Proxy
subclasses. Print the value of the pointer field and then the class
name. This will undoubtedly need improving, but on this machine,
anyway, it matches the numeric value resulting from calling

	printf("%ul", (long) object)

on the JNI side. Talk about a hack, I know, but it was useful for
debugging the toggle references.