Replace GObject.ref() and unRef() with add/removeToggleRef()
Switching to the full blown reference management system means taking
ToggleRefs instead of a normal GObject ref or not taking one. Although
the use of g_object_add_toggle_ref() is an implementation matter, we
might as well be consistent and call the methods we expose in the
translation layer what they purport to wrap.

So rename GObject.ref() to GObject.addToggleRef() and GObject.unRef()
to GObject.removeToggleRef().  

Note that the native implementation actually calls new code in
forthcoming bindings_java_memory.c to carry out the heavy lifting.

Slightly modify the debug output in the Proxy subclass constructors to
tab align with other debug prints.