Delay dropping initial GObject Ref to avoid ping pong effect
Although logically correct, the code we had before resulted in the
ToggleRef being the last ref count to a GtkWidget. The strange result
was that when setters are called on the Widget prior to its being
added to a Container, their ref count is incremented in the GTK
functions implementation, then decremented at the end. This meant we
had a strong/weak/strong/weak/strong/weak/strong.... sequence of Java
reference creation and removal. It looked a bit off and made debugging
tougher because of the noise.

So delay the necessary g_object_unref() call to discard the original
(probably floating) Ref by stuffing it into a GIdle callback,
internally named bindings_java_memory_deref().

By delaying the ref count reduction, the inital strong Java reference
is left alone, but if it's going to shift the Java reference to be
weak, it happens very quickly (when the main loop next [first] fires).