Update unit tests for refactored method names
The original ValidateReferenceCounting merely checked to make sure
that nothing blew up if you decremented and restored a Ref to a
Widget already packed into a Container. Refactored, it removes and
restores the ToggleRef and still passes. Theoretically that is a good
thing, but these tests are not conclusive.

Needing to add a change to GValue.c adding an additional check before
calling g_object_unref() if it thinks it has a GObject may indicate a
leak somewhere, but I can't find it. For now, this avoids a
GLib-GObject-CRITICAL in one unit test.

Request one final garbage collection before exiting UnitTests.
Theoretically, there shouldn't be anything left to collect, but one
last check post test cases.