Add improves and some new tests to validate memory management.
A pair of tests to check that no denaturation occurs and that the object
retrieved from Gtk+ wrapped methods (such as getChild()...) is the same
that original object have been implemented.

Also it has been taken into account the new way of removing Java 
references to GObject by delaying it a bit with a g_iddle_add.

At current time, all test are passed except for the one that checks that 
cyclic references are managed in the right way. This issue will need 
some changes in the way signal handlers are managed at this time.

The memory management test still needs some improves to test also what 
things are being done on Gtk+ side...