Move strong reference to Signal handler back to Java side
We ended up with a circular reference when we had a object hooking up
a signal that was itself (self-delegattion).

The solution proposed by Vreixo Formoso was to move the strong Java
reference from the JNI side to the Java side so that the garbage
collector can detect it as a strong reference from an object which
becomes no-longer strongly reachable.

So add a private HashSet in org.gnome.glib.Object to do just that.
When connecting a signal, quickly call addHandler() to store it in the
Set. Then change the reference in our BindingsJavaClosure [as set by
bindings_java_closure_new() in src/jni/bindings_java_signal.c] to a
weak one.