Add mockups for GtkMisc's setAlignment() and setPadding()
Quickly implement mockup code for two methods in GtkMisc so we can add
them to our public Misc API class as a bugfix during presentation at
Mukthi 7.04 conference.

Change the packing of the Label in Experiment so that space is
allocated to the widget allowing us to test the above code when
implemented.