Import basic defs files parser from pygtk
As a first step in working with the .defs data in common with the
Python bindings project, we need some tools to manipulate these files.
Although our code generator will not be based on this code,
nevertheless using the pygtk defsparser allows us to roundtrip
individual defs file blocks.

This code is licenced under the LGPL and is copyright by the people
listed in the AUTHORS file and code history of the respective files
that are impored here.