Modify pygtk code generator to merely splice the defs data stream
In order to achieve our goal of interoperation with pygtk, we need to
work from the same .defs data. Their data, however, is monolithic and
we wish it in individual files, one .defs file per class.

Therefore hack their code generator to do a round trip read/write and
append each define block to the appropriate individual file. Blocks
that we don't know what to do with go to stdout.