Add the .defs file with all the broken blocks
Add src/util/missing.defs, containing the blocks rejected from being
allocated to a specific class when demux ran. To fix these we'll have
to fix the upstream .defs files fed into the demuxer.

This isn't the best place for this to sit, but since everything *.defs
in src/defs/ will be processed by the code generator, better to keep
it out of that tree for now.